+++ src/main/resources/templates/result.html
@@ -67,7 +67,7 @@ body {
             <br/>
           {{/each}}
           <br />
-          <div>{{#each result.ranking}}{{this.name}}{{/each}}이 우승했습니다.</div>
+          <div>{{#each result.winner}}{{this.name}}{{/each}}이 우승했습니다.</div>
         </div>
       </div>
     </div>
